HSEARCH-4678 Move `UuidGenerationStrategy` to config package and mark as incubating
